style: index organized
