+++ public/css/styles.css
@@ -214,7 +214,7 @@ img {
     display: flex;
     justify-content: flex-end;
     align-items: center;
-    gap: 20px;
+    gap: 30px;
     padding: 15px 20px;
     background-color: #003153;
     min-height: 50px;
